+++ packages/canon/package.json
@@ -45,6 +45,7 @@
     "@remixicon/react": "^4.6.0",
     "@tanstack/react-table": "^8.21.3",
     "clsx": "^2.1.1",
+    "motion": "^12.20.1",
     "react-aria-components": "^1.10.1"
   },
   "devDependencies": {
